backend/requirements.txt

six==1.16.0
tg-logger==3.3
urllib3==1.26.9
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
gevent==21.12.0
google-auth==2.8.0
google-auth-oauthlib==0.5.2
greenlet==1.1.2
gspread==5.4.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyTelegramBotAPI==4.5.1
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tg-logger==3.3
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
